:test_tube: :sparkles: add more convenient test structure
